+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Removed `RestartableTimer`. `package:async` already provides such a class.
 * Added a `timeAsyncOperation` function as an asynchronous version of
   `timeOperation`.
+* Added a `parseOptionValue` extension method on `package:args`'s `ArgResults`.
 
 ## 0.3.0